remove ViewItemHandler
