+++ miru/abc/item_handler.py
@@ -337,20 +337,6 @@ class ItemHandler(Sequence, abc.ABC, t.Generic[T]):  # type: ignore[type-arg]
         return action_rows
 
 
-class ViewItemHandler(ItemHandler[hikari.impl.ActionRowBuilder]):
-    def build(self) -> t.Sequence[hikari.impl.ActionRowBuilder]:
-        """Creates the action rows the item handler represents.
-
-        Returns
-        -------
-        List[hikari.impl.ActionRowBuilder]
-            A list of action rows containing all items attached to this item handler,
-            converted to hikari component objects. If the item handler has no items attached,
-            this returns an empty list.
-        """
-        return self._build_inner(hikari.impl.ActionRowBuilder)
-
-
 # MIT License
 #
 # Copyright (c) 2022-present HyperGH
